backend/requirements.txt
flask
flask-sqlalchemy
flask-bcrypt
flask-login
flask-cors
flask-talisman
flask-limiter
flask-mail
itsdangerous
tensorflow>=2.0
keras
numpy
opencv-python
werkzeug
python-dotenv
psycopg2-binary
Pillow
torch
torchvision
timm
scikit-learn
alembic
requests
pandas
apscheduler>=3.9.1